requirements.txt
discord.py==2.4.0
yt-dlp==2025.1.26
PyNaCl==1.5.0
youtube-search==2.1.2